Trigger fixes

+++ database/migrations/1550233134621_delete_comment_action_schema.js
@@ -9,9 +9,7 @@ class DeleteCommentActionSchema extends Schema {
         await Database
             .raw(
                 'CREATE TRIGGER `delete_comment_action` BEFORE DELETE ON `comments`' +
-                'FOR EACH ROW DELETE FROM actions' +
-                'WHERE receiver_id = (SELECT owner_id FROM posts WHERE OLD.post_id)' +
-                'AND initiator_id=OLD.owner_id AND type=2 AND entity_id=OLD.id'
+                'FOR EACH ROW DELETE FROM actions WHERE receiver_id = (SELECT owner_id FROM posts WHERE OLD.post_id) AND initiator_id=OLD.owner_id AND type=2 AND entity_id=OLD.id'
             );
     }
